basic sanitizer working
