+++ internal/article/usecase/articleUseCase.go
@@ -5,8 +5,6 @@ import (
 
 	"context"
 
-	"github.com/microcosm-cc/bluemonday"
-
 	repository "github.com/go-park-mail-ru/2021_2_SaberDevs/internal/article/repository"
 	"github.com/jmoiron/sqlx"
 
@@ -15,25 +13,6 @@ import (
 
 type articleUseCase struct {
 	articleRepo amodels.ArticleRepository
-}
-
-func SanitizeArticle(a *amodels.Article) *amodels.Article {
-	s := bluemonday.StrictPolicy()
-	l := bluemonday.UGCPolicy()
-	a.AuthorAvatar = s.Sanitize(a.AuthorAvatar)
-	a.AuthorName = s.Sanitize(a.AuthorName)
-	a.AuthorUrl = s.Sanitize(a.AuthorUrl)
-	//a.Comments = s.Sanitize(a.Comments) //not a string
-	a.CommentsUrl = s.Sanitize(a.CommentsUrl)
-	a.Id = s.Sanitize(a.Id)
-	// a.Likes = s.Sanitize(a.Likes)//not a string
-	a.PreviewUrl = s.Sanitize(a.PreviewUrl)
-	for i := range a.Tags {
-		a.Tags[i] = s.Sanitize(a.Tags[i])
-	}
-	a.Text = l.Sanitize(a.Text)
-	a.Title = s.Sanitize(a.Title)
-	return a
 }
 
 func NewArticleUsecase(db *sqlx.DB) amodels.ArticleUseCase {
@@ -73,7 +52,6 @@ func (m *articleUseCase) GetByAuthor(ctx context.Context, author string) (result
 }
 
 func (m *articleUseCase) Store(ctx context.Context, a *amodels.Article) error {
-	a = SanitizeArticle(a)
 	err := m.articleRepo.Store(ctx, a)
 	return err
 }
